Add "Jump to Target Article" action
This new action is particularly useful when the "Automatically scroll to
target article" option is disabled. Even if the user prefers
translations from higher-priority dictionaries most of the time,
sometimes he may want to read the target article. With this action it is
possible to scroll to it via the new action's Alt+T shortcut.