fix: respect the setting of QT_ENABLE_HIGHDPI_SCALING of the system (#2077)
* fix: respect the setting of QT_ENABLE_HIGHDPI_SCALING of the system

Setting QT_ENABLE_HIGHDPI_SCALING to 1 might not work well in some
cases. Instead of hard-coding it to 1, this commit checks if it is
already set, and if it is empty or not set, then set it to 1. This
allows overriding the QT_ENABLE_HIGHDPI_SCALING to 0 by respecting the
system-wide setting of environment variables.

* fix: move setHighDpiScaleFactorRoundingPolicy out of the check of QT_ENABLE_HIGHDPI_SCALING

* [autofix.ci] apply automated fixes

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>